+++ .gitignore
@@ -1,0 +1,18 @@
+# Created by .ignore support plugin (hsz.mobi)
+### Java template
+*.class
+
+out/
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.ear
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
+
